+++ temporal_random_walk/src/core/temporal_random_walk_kernels_step_based.cuh
@@ -19,12 +19,12 @@ namespace temporal_random_walk {
         const int max_walk_len,
         const size_t num_walks,
         const double *__restrict__ rand_nums) {
-        const int walk_idx = blockIdx.x * blockDim.x + threadIdx.x;
+        const size_t walk_idx = blockIdx.x * blockDim.x + threadIdx.x;
         if (walk_idx >= num_walks) return;
 
-        const int rand_nums_start_offset =
-            walk_idx +                          // To account extra value in all previous walk's start pickers.
-                (walk_idx * max_walk_len * 2);  // To account all 2 rand numbers for all other steps in the previous walks.
+        const size_t rand_nums_start_offset =
+            static_cast<size_t>(walk_idx) +                                           // To account extra value in all previous walk's start pickers.
+            (static_cast<size_t>(walk_idx) * static_cast<size_t>(max_walk_len) * 2);  // To account all 2 rand numbers for all other steps in the previous walks.
 
         Edge start_edge;
         if (start_node_ids[walk_idx] == -1) {
@@ -80,21 +80,21 @@ namespace temporal_random_walk {
         const size_t num_walks,
         const double *__restrict__ rand_nums) {
 
-        const int walk_idx = blockIdx.x * blockDim.x + threadIdx.x;
+        const size_t walk_idx = blockIdx.x * blockDim.x + threadIdx.x;
         if (walk_idx >= num_walks) return;
 
         if (step_number >= max_walk_len - 1) {
             return;
         }
 
-        const size_t offset = walk_idx * max_walk_len + step_number; // Get endpoint of previous step (step_number - 1). And endpoint is (step_number - 1 + 1).
+        const size_t offset = static_cast<size_t>(walk_idx) * static_cast<size_t>(max_walk_len) + static_cast<size_t>(step_number); // Get endpoint of previous step (step_number - 1). And endpoint is (step_number - 1 + 1).
         const int last_node = walk_set->nodes[offset];
         const int last_ts = walk_set->timestamps[offset];
 
-        const int rand_nums_start_offset =
-            walk_idx +                              // To account extra value in all previous walk's start pickers.
-                (walk_idx * max_walk_len * 2) +     // To account all 2 rand numbers for all other steps in the previous walks.
-                    (step_number * 2 + 1);          // To account for random numbers used in the current walk.
+        const size_t rand_nums_start_offset =
+            static_cast<size_t>(walk_idx) +                                               // To account extra value in all previous walk's start pickers.
+            (static_cast<size_t>(walk_idx) * static_cast<size_t>(max_walk_len) * 2) +     // To account all 2 rand numbers for all other steps in the previous walks.
+            (static_cast<size_t>(step_number) * 2 + 1);                                   // To account for random numbers used in the current walk.
 
         const Edge next_edge = temporal_graph::get_node_edge_at_device<Forward, EdgePickerType, IsDirected>(
                 temporal_graph,
@@ -116,7 +116,7 @@ namespace temporal_random_walk {
     }
 
     __global__ static void reverse_walks_kernel(const WalkSet *__restrict__ walk_set, const size_t num_walks) {
-        const int walk_idx = blockIdx.x * blockDim.x + threadIdx.x;
+        const size_t walk_idx = blockIdx.x * blockDim.x + threadIdx.x;
         if (walk_idx >= num_walks) return;
         walk_set->reverse_walk(walk_idx);
     }
